# openapi.json
{"openapi":"3.0.1","info":{"title":"Mini Leave Management System API","description":"A comprehensive REST API for managing employee leave requests","contact":{"name":"Avi","url":"https://github.com/Avi1606","email":"avi@example.com"},"license":{"name":"MIT License","url":"https://opensource.org/licenses/MIT"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Employee Management","description":"APIs for managing employees"},{"name":"Leave Request Management","description":"APIs for managing leave requests"}],"paths":{"/api/leave-requests/{id}/reject":{"put":{"tags":["Leave Request Management"],"summary":"Reject leave request","description":"Reject a pending leave request","operationId":"rejectLeaveRequest","parameters":[{"name":"id","in":"path","description":"Leave request ID","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"},"description":"Rejection data"}}},"required":true},"responses":{"200":{"description":"Leave request rejected successfully","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Invalid request or business rule violation","content":{"*/*":{"schema":{"type":"object"}}}},"404":{"description":"Leave request not found","content":{"*/*":{"schema":{"type":"object"}}}},"403":{"description":"Manager does not have authority to reject","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/leave-requests/{id}/approve":{"put":{"tags":["Leave Request Management"],"summary":"Approve leave request","description":"Approve a pending leave request","operationId":"approveLeaveRequest","parameters":[{"name":"id","in":"path","description":"Leave request ID","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"},"description":"Approval data"}}},"required":true},"responses":{"403":{"description":"Manager does not have authority to approve","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Invalid request or business rule violation","content":{"*/*":{"schema":{"type":"object"}}}},"404":{"description":"Leave request not found","content":{"*/*":{"schema":{"type":"object"}}}},"200":{"description":"Leave request approved successfully","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/employees/{id}":{"get":{"tags":["Employee Management"],"summary":"Get employee by ID","description":"Retrieve a specific employee by their ID","operationId":"getEmployeeById","parameters":[{"name":"id","in":"path","description":"Employee ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Employee not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeDTO"}}}},"200":{"description":"Employee found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeDTO"}}}}}},"put":{"tags":["Employee Management"],"summary":"Update employee","description":"Update an existing employee record","operationId":"updateEmployee","parameters":[{"name":"id","in":"path","description":"Employee ID","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeDTO"}}},"required":true},"responses":{"404":{"description":"Employee not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeDTO"}}}},"409":{"description":"Email already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeDTO"}}}},"200":{"description":"Employee updated successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeDTO"}}}},"400":{"description":"Invalid input data","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeDTO"}}}}}},"delete":{"tags":["Employee Management"],"summary":"Delete employee","description":"Delete an employee record","operationId":"deleteEmployee","parameters":[{"name":"id","in":"path","description":"Employee ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Employee not found"},"204":{"description":"Employee deleted successfully"}}}},"/api/leave-requests":{"get":{"tags":["Leave Request Management"],"summary":"Get all leave requests","description":"Retrieve a list of all leave requests","operationId":"getAllLeaveRequests","responses":{"200":{"description":"Successfully retrieved leave requests","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LeaveRequestDTO"}}}}}}},"post":{"tags":["Leave Request Management"],"summary":"Submit leave request","description":"Submit a new leave request","operationId":"submitLeaveRequest","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeaveRequestDTO"}}},"required":true},"responses":{"404":{"description":"Employee not found","content":{"*/*":{"schema":{"type":"object"}}}},"201":{"description":"Leave request submitted successfully","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Invalid input data or business rule violation","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/employees":{"get":{"tags":["Employee Management"],"summary":"Get all employees","description":"Retrieve a list of all employees","operationId":"getAllEmployees","responses":{"200":{"description":"Successfully retrieved employees","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EmployeeDTO"}}}}}}},"post":{"tags":["Employee Management"],"summary":"Create new employee","description":"Create a new employee record","operationId":"createEmployee","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeDTO"}}},"required":true},"responses":{"409":{"description":"Employee with email already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeDTO"}}}},"400":{"description":"Invalid input data","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeDTO"}}}},"201":{"description":"Employee created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeDTO"}}}}}}},"/api/leave-requests/{id}":{"get":{"tags":["Leave Request Management"],"summary":"Get leave request by ID","description":"Retrieve a specific leave request by its ID","operationId":"getLeaveRequestById","parameters":[{"name":"id","in":"path","description":"Leave request ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Leave request found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LeaveRequestDTO"}}}},"404":{"description":"Leave request not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LeaveRequestDTO"}}}}}}},"/api/leave-requests/manager/{managerId}":{"get":{"tags":["Leave Request Management"],"summary":"Get leave requests for manager","description":"Retrieve leave requests for team members of a specific manager","operationId":"getLeaveRequestsForManager","parameters":[{"name":"managerId","in":"path","description":"Manager ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LeaveRequestDTO"}}}}}}}},"/api/leave-requests/employee/{employeeId}":{"get":{"tags":["Leave Request Management"],"summary":"Get leave requests by employee","description":"Retrieve leave requests for a specific employee","operationId":"getLeaveRequestsByEmployee","parameters":[{"name":"employeeId","in":"path","description":"Employee ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LeaveRequestDTO"}}}}}}}},"/api/employees/team/{managerId}":{"get":{"tags":["Employee Management"],"summary":"Get team members","description":"Retrieve team members for a specific manager","operationId":"getTeamMembers","parameters":[{"name":"managerId","in":"path","description":"Manager ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EmployeeDTO"}}}}}}}},"/api/employees/department/{department}":{"get":{"tags":["Employee Management"],"summary":"Get employees by department","description":"Retrieve employees from a specific department","operationId":"getEmployeesByDepartment","parameters":[{"name":"department","in":"path","description":"Department name","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EmployeeDTO"}}}}}}}}},"components":{"schemas":{"EmployeeDTO":{"required":["department","email","name","role"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"maxLength":100,"minLength":2,"type":"string"},"email":{"type":"string"},"department":{"maxLength":50,"minLength":2,"type":"string"},"role":{"maxLength":50,"minLength":2,"type":"string"},"managerId":{"type":"integer","format":"int64"},"managerName":{"type":"string"}},"description":"Employee data"},"LeaveRequestDTO":{"required":["employeeId","endDate","leaveType","reason","startDate"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"employeeId":{"type":"integer","format":"int64"},"employeeName":{"type":"string"},"leaveType":{"type":"string","enum":["SICK","CASUAL","ANNUAL"]},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"reason":{"maxLength":500,"minLength":10,"type":"string"},"status":{"type":"string","enum":["PENDING","APPROVED","REJECTED"]},"appliedDate":{"type":"string","format":"date"},"approvedBy":{"type":"integer","format":"int64"},"approvedByName":{"type":"string"},"approvedDate":{"type":"string","format":"date"},"comments":{"maxLength":255,"minLength":0,"type":"string"},"workingDays":{"type":"integer","format":"int64"}},"description":"Leave request data"}}}}